# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdminActionResponse": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "AdminActionResponse",
        "type": "object"
      },
...
      "AdminPaperListResponse": {
        "properties": {
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/AdminPaperListMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "AdminPaperListResponse",
        "type": "object"
      },
      "AdminUserResponse": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "role"
        ],
        "title": "AdminUserResponse",
        "type": "object"
      },
      "AnswerUpsertRequest": {
        "additionalProperties": false,
        "properties": {
          "content_language": {
            "default": "si",
            "enum": [
              "en",
              "si",
              "ta"
            ],
            "title": "Content Language",
            "type": "string"
          },
          "explanation_markdown": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation Markdown"
          },
          "solution_markdown": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Solution Markdown",
            "type": "string"
          }
        },
        "required": [
          "solution_markdown"
        ],
        "title": "AnswerUpsertRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaperCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "exam_year_id": {
            "format": "uuid",
            "title": "Exam Year Id",
            "type": "string"
          },
          "level_id": {
            "format": "uuid",
            "title": "Level Id",
            "type": "string"
          },
          "medium": {
            "default": "si",
            "enum": [
              "en",
              "si",
              "ta"
            ],
            "title": "Medium",
            "type": "string"
          },
          "paper_number": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Paper Number",
            "type": "string"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "total_marks": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,9}0*$)\\d{0,6}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Marks"
          }
        },
        "required": [
          "title",
          "level_id",
          "subject_id",
          "exam_year_id",
          "paper_number"
        ],
        "title": "PaperCreateRequest",
        "type": "object"
      },
...
      "PaperListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PaperSummary"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PaperListMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PaperListResponse",
        "type": "object"
      },
      "PaperPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "medium": {
            "anyOf": [
              {
                "enum": [
                  "en",
                  "si",
                  "ta"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medium"
          },
          "paper_number": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Number"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "total_marks": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,9}0*$)\\d{0,6}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Marks"
          }
        },
        "title": "PaperPatchRequest",
        "type": "object"
      },
...
      "QuestionCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "marks": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,6}|(?=[\\d.]{1,9}0*$)\\d{0,6}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Marks"
          },
          "number_label": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Number Label",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "prompt_markdown": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Prompt Markdown",
            "type": "string"
          }
        },
        "required": [
          "number_label",
          "prompt_markdown",
          "marks"
        ],
        "title": "QuestionCreateRequest",
        "type": "object"
      },
...
      "VideoSourceCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "is_primary": {
            "default": false,
            "title": "Is Primary",
            "type": "boolean"
          },
          "original_url": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Original Url",
            "type": "string"
          },
          "provider": {
            "enum": [
              "youtube",
              "tiktok",
              "facebook"
            ],
            "title": "Provider",
            "type": "string"
          },
          "question_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          }
        },
        "required": [
          "provider",
          "original_url"
        ],
        "title": "VideoSourceCreateRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for O/L Mathematics papers, answers, marking schemes, and video links.",
    "title": "Siyana Maths API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/me": {
      "get": {
        "operationId": "current_admin_api_v1_admin_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Current Admin",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/papers": {
      "get": {
        "operationId": "list_admin_papers_api_v1_admin_papers_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(draft|published|archived)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPaperListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List papers for the admin workspace",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_paper_api_v1_admin_papers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a draft O/L Mathematics paper",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/papers/{paper_id}": {
      "get": {
        "operationId": "get_admin_paper_api_v1_admin_papers__paper_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Paper Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get a complete editable paper",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "update_paper_api_v1_admin_papers__paper_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Paper Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update paper metadata",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/papers/{paper_id}/publish": {
      "post": {
        "description": "Publishes only when every question has an answer and marking scheme, mark totals are valid, and at least one video source exists.",
        "operationId": "publish_paper_api_v1_admin_papers__paper_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Paper Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate and publish a paper",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/papers/{paper_id}/questions": {
      "post": {
        "operationId": "create_question_api_v1_admin_papers__paper_id__questions_post",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Paper Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add an ordered question to a draft paper",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/papers/{paper_id}/video-sources": {
      "post": {
        "operationId": "create_video_source_api_v1_admin_papers__paper_id__video_sources_post",
        "parameters": [
          {
            "in": "path",
            "name": "paper_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Paper Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoSourceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Attach a paper- or question-level video source",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/questions/{question_id}/answer": {
      "put": {
        "operationId": "save_answer_api_v1_admin_questions__question_id__answer_put",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create or replace a Sinhala answer",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/questions/{question_id}/marking-scheme": {
      "post": {
        "operationId": "create_marking_item_api_v1_admin_questions__question_id__marking_scheme_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkingItemCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add a full marking-scheme item",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/papers": {
      "get": {
        "operationId": "list_papers_api_v1_papers_get",
        "parameters": [
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2200,
                  "minimum": 1900,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "in": "query",
            "name": "paper_number",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Paper Number"
            }
          },
          {
            "in": "query",
            "name": "medium",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(en|si|ta)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Medium"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Papers",
        "tags": [
          "papers"
        ]
      }
    },
...
  },
...
}